Split engine/shared/engine.cpp
