+++ CMakeLists.txt
@@ -1760,6 +1760,7 @@ set_src(ENGINE_SHARED GLOB_RECURSE src/engine/shared
   econ.cpp
   econ.h
   engine.cpp
+  engine.h
   fifo.cpp
   fifo.h
   filecollection.cpp
